Add customised Bootstrap style.
This is a lot like the default Bootstrap style, except that the primary
brand colour has been set to #1d88c9.  The inverse nav bar background
has also been set to this colour, and the text colour for the inverse
nav bar is made lighter than the default to compensate.