Split the site header and footer into separate files.
This is so we don't have to copy the repeated parts of the site into
every page's template, which would be a real pain in the butt if we ever
wanted to change anything.

Each page now needs to include header.php at the beginning (just after
including the CouchCMS library) and footer.php at the end (just before
calling COUCH::invoke().

It's still a bit more boilerplatey than I would like, but you can't get
much better than this in a PHP setup.